broke out measurementCard => addMeasurement components
