+++ client/components/addMeasurement.js
@@ -16,33 +16,16 @@ import FormControlLabel from '@material-ui/core/FormControlLabel'
 import Select from '@material-ui/core/Select'
 import {withStyles, withTheme} from '@material-ui/core/styles'
 
-const styles = theme => ({
-  root: {
-    display: 'flex',
-    flexWrap: 'wrap',
-    color: 'white',
-    fontWeight: '600',
-    fontFamily: 'CircularStd, sans-serif',
-    letterSpacing: '0.1rem'
-  },
-  formControl: {
-    margin: theme.spacing.unit,
-    minWidth: 120
-  },
-  selectEmpty: {
-    marginTop: theme.spacing.unit * 2
-  }
-})
-
 class AddMeasurement extends Component {
+  constructor() {
+    super()
+    this.state = {
+      category: '',
+      data: 0
+    }
+  }
   state = {
     type: ''
-  }
-
-  componentDidMount() {
-    this.setState({
-      // labelWidth: ReactDOM.findDOMNode(this.InputLabelRef).offsetWidth
-    })
   }
 
   handleChange = event => {
@@ -53,83 +36,37 @@ class AddMeasurement extends Component {
     const {classes} = this.props
 
     return (
-      <Card id="addMeasurement">
-        <div id="addMeasurementHeader">
-          <h3>+ ADD MEASUREMENT</h3>
-        </div>
+      <form className={classes.form}>
         <CardContent>
-          <form className={classes.form}>
-            <FormControl margin="normal" fullWidth>
-              <InputLabel htmlFor="type">measurement type</InputLabel>
-              <Select
-                value={this.state.type}
-                onChange={this.handleChange}
-                name="type"
-              >
-                <MenuItem value="weight">weight</MenuItem>
-                <MenuItem value="body fat">body fat percentage</MenuItem>
-                <MenuItem value="waist">waist</MenuItem>
-                <MenuItem value="hips">hips</MenuItem>
-                <MenuItem value="thighs">thighs</MenuItem>
-                <MenuItem value="chest">chest</MenuItem>
-                <MenuItem value="arms">arms</MenuItem>
-              </Select>
-            </FormControl>
-            <FormControl margin="normal" fullWidth>
-              <InputLabel htmlFor="data">enter amount</InputLabel>
-              <Input name="data" id="data" />
-            </FormControl>
-          </form>
+          <FormControl margin="normal" fullWidth>
+            <InputLabel htmlFor="type">measurement type</InputLabel>
+            <Select
+              value={this.state.type}
+              onChange={this.handleChange}
+              name="type"
+            >
+              <MenuItem value="weight">weight</MenuItem>
+              <MenuItem value="body fat">body fat percentage</MenuItem>
+              <MenuItem value="waist">waist</MenuItem>
+              <MenuItem value="hips">hips</MenuItem>
+              <MenuItem value="thighs">thighs</MenuItem>
+              <MenuItem value="chest">chest</MenuItem>
+              <MenuItem value="arms">arms</MenuItem>
+            </Select>
+          </FormControl>
+          <FormControl margin="normal" fullWidth>
+            <InputLabel htmlFor="data">enter amount</InputLabel>
+            <Input name="data" id="data" />
+          </FormControl>
         </CardContent>
         <CardActions>
           <Button id="addMeasurementSubmit" fullWidth>
             SUBMIT
           </Button>
         </CardActions>
-      </Card>
+      </form>
     )
   }
 }
 
-export default withStyles(styles)(AddMeasurement)
-
-// <div>
-// <Grid container spacing={8} alignItems="flex-end">
-//             <Grid item>
-//               <h5>TYPE</h5>
-//             </Grid>
-
-//             {/* ========= TYPE ========= */}
-//             <Grid item>
-//               <FormControl fullWidth>
-//                 {/* <InputLabel htmlFor="type">measurement type</InputLabel> */}
-//                 <Select
-//                   value={this.state.type}
-//                   onChange={this.handleChange}
-//                   name="type"
-//                 >
-//                   <MenuItem value="weight">Weight</MenuItem>
-//                   <MenuItem value="body fat">Body Fat Percentage</MenuItem>
-//                   <MenuItem value="waist">Waist</MenuItem>
-//                   <MenuItem value="hips">Hips</MenuItem>
-//                   <MenuItem value="thighs">Thighs</MenuItem>
-//                   <MenuItem value="chest">Chest</MenuItem>
-//                   <MenuItem value="arms">Arms</MenuItem>
-//                 </Select>
-//               </FormControl>
-//             </Grid>
-//           </Grid>
-//           {/* ========= DATA ========= */}
-//           <Grid container spacing={8} alignItems="flex-end">
-//             <Grid item>
-//               <h5>DATA</h5>
-//             </Grid>
-//             <Grid item>
-//               <TextField
-//                 id="input-with-icon-grid"
-//                 label="enter amount"
-//                 fullWidth
-//               />
-//             </Grid>
-//           </Grid>
-//           </div>
+export default AddMeasurement
