+++ client/components/fitness.js
@@ -2,7 +2,7 @@ import React from 'react'
 import {connect} from 'react-redux'
 import CssBaseline from '@material-ui/core/CssBaseline'
 import SimpleLineChart from './simpleLineChart'
-import AddMeasurement from './addMeasurement'
+import MeasurementCard from './measurementCard'
 
 const Fitness = props => {
   const {user} = props
@@ -17,7 +17,7 @@ const Fitness = props => {
         <div className="chartContainer">
           <SimpleLineChart />
         </div>
-        <AddMeasurement className="fitnessRight" />
+        <MeasurementCard className="fitnessRight" />
       </div>
       {/* <h2>dailies</h2>
       <div id="dailiesContainer">
