+++ client/components/index.js
@@ -15,6 +15,7 @@ export {default as Dashboard} from './dashboard'
 export {default as SimpleLineChart} from './simpleLineChart'
 export {default as SimplePieChart} from './simplePieChart'
 export {default as PersistentDrawer} from './persistentDrawer'
+export {default as MeasurementCard} from './measurementCard'
 export {default as AddMeasurement} from './addMeasurement'
 export {default as FoodCard} from './foodCard'
 export {default as FoodItem} from './foodItem'
